backend/app/services/search_tools.py
import os
import requests
from typing import List, Dict, Any
from duckduckgo_search import DDGS
import asyncio
import wikipedia
from dotenv import load_dotenv
import urllib3

# 禁用SSL警告
urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)

load_dotenv()

class SearchTools:
    def __init__(self):
        self.serpapi_key = os.getenv("SERPAPI_API_KEY")
    
    async def google_search(self, query: str, num_results: int = 10) -> List[Dict[str, Any]]:
        """使用SerpAPI进行Google搜索"""
        if not self.serpapi_key:
            # 如果没有SerpAPI密钥，使用DuckDuckGo作为替代
            return await self.duckduckgo_search(query, num_results)
        
        url = "https://serpapi.com/search"
        params = {
            "q": query,
            "api_key": self.serpapi_key,
            "num": num_results,
            "engine": "google"
        }
        
        try:
            response = requests.get(url, params=params, timeout=10, verify=False)
            if response.status_code == 200:
                data = response.json()
                results = []
                
                for result in data.get("organic_results", []):
                    results.append({
                        "title": result.get("title", ""),
                        "link": result.get("link", ""),
                        "snippet": result.get("snippet", ""),
                        "source": "google"
                    })
                
                return results
            else:
                # 如果Google搜索失败，回退到DuckDuckGo
                return await self.duckduckgo_search(query, num_results)
        except Exception as e:
            print(f"Google search error: {e}")
            return await self.duckduckgo_search(query, num_results)
    
    async def duckduckgo_search(self, query: str, num_results: int = 10) -> List[Dict[str, Any]]:
        """使用DuckDuckGo搜索"""
        try:
            # DuckDuckGo 4.3.0版本的正确用法
            ddgs = DDGS()
            results = []
            # 使用迭代器获取结果
            search_results = list(ddgs.text(query, max_results=num_results))
            for result in search_results:
                results.append({
                    "title": result.get("title", ""),
                    "link": result.get("href", ""),
                    "snippet": result.get("body", ""),
                    "source": "duckduckgo"
                })
            return results
        except Exception as e:
            print(f"DuckDuckGo search error: {e}")
            # 返回一些模拟结果以保证系统正常运行
            return [{
                "title": f"搜索结果: {query}",
                "link": "https://example.com",
                "snippet": f"关于'{query}'的相关信息。",
                "source": "duckduckgo"
            }]
    
    async def wikipedia_search(self, query: str, num_results: int = 5) -> List[Dict[str, Any]]:
        """使用Wikipedia搜索"""
        try:
            # 设置中文Wikipedia和超时
            wikipedia.set_lang("zh")
            
            # 搜索页面，增加错误处理
            search_results = wikipedia.search(query, results=num_results)
            results = []
            
            for title in search_results:
                try:
                    page = wikipedia.page(title)
                    results.append({
                        "title": page.title,
                        "link": page.url,
                        "snippet": page.summary[:300] + "..." if len(page.summary) > 300 else page.summary,
                        "source": "wikipedia"
                    })
                except wikipedia.exceptions.DisambiguationError as e:
                    # 如果有歧义，使用第一个选项
                    if e.options:
                        try:
                            page = wikipedia.page(e.options[0])
                            results.append({
                                "title": page.title,
                                "link": page.url,
                                "snippet": page.summary[:300] + "..." if len(page.summary) > 300 else page.summary,
                                "source": "wikipedia"
                            })
                        except:
                            continue
                except:
                    continue
            
            return results
        except Exception as e:
            print(f"Wikipedia search error: {e}")
            return []
    
    async def academic_search(self, query: str, num_results: int = 5) -> List[Dict[str, Any]]:
        """学术搜索（可以后续集成Semantic Scholar API等）"""
        # 目前使用Google学术搜索的变体
        academic_query = f"site:scholar.google.com OR site:arxiv.org OR site:researchgate.net {query}"
        return await self.google_search(academic_query, num_results)
    
    async def comprehensive_search(self, query: str) -> Dict[str, List[Dict[str, Any]]]:
        """综合搜索，使用多个搜索引擎"""
        results = {}
        
        # Google/DuckDuckGo 搜索
        results["web"] = await self.google_search(query, 8)
        
        # Wikipedia 搜索
        results["wikipedia"] = await self.wikipedia_search(query, 3)
        
        # 学术搜索
        results["academic"] = await self.academic_search(query, 3)
        
        return results

# 全局实例
search_tools = SearchTools()