backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
duckduckgo-search==4.3.0
google-search-results==2.4.2
wikipedia==1.4.0
sqlalchemy==2.0.23
python-multipart==0.0.6